iss: don't perform bounds check during integer cast in write_register
The write_register function prototype must satisfy the function
declaration from the iss_syscall_if type. For this reason, both the RV32
as well as the RV64 implementation use a uint64_t type for the register
value parameter. Without this commit, this value is converted to a
uint32_t in the RV32 implementation using a Boost lexical_cast. However,
lexical_cast performs a runtime bounds check. This bounds check can fail
if a signed negative integer is passed to the write_register function
since this value would be promoted implicitly to a uint64_t
(e.g. -1 would become (2^64)-1). For example, this happens in
the execute_syscall implementation since some system calls
return -1 to indicate a failure condition. Presently, this causes
sw/sys-read to not work correctly.

This commit fixes this issue by performing a cast without integer
boundary checks in write_register.