GUI-VP: Added sifive_test module which allows stopping the simulation from within
Inspired by qemu-riscv.

Writing specific values to specific adresses causes a halt (or reboot).
This allows the firmware (linux) to stop the simulation from within.

The interface would also support reboot, but since reboot is not
implemented in the vp yet, we also simply stop the simulation in this
case.

Regarding compatibility:
The sifive,test module is not documented for the fu540. However it does
also not conflict with other documented modules in the fu540 memory map
(https://static.dev.sifive.com/FU540-C000-v1.0.pdf).

Needed in devicetree:
{{{
+  poweroff {
+    value = <0x5555>;
+    offset = <0x00>;
+    regmap = <&test>;
+    compatible = "syscon-poweroff";
+  };
+
+  reboot {
+    value = <0x7777>;
+    offset = <0x00>;
+    regmap = <&test>;
+    compatible = "syscon-reboot";
+  };
+
+  test: test@100000 {
+    reg = <0x100000 0x1000>;
+    compatible = "sifive,test1", "sifive,test0", "syscon";
+  };
}}}