syscall: Respect exit code specified by the executed software
The SYS_exit system call takes an exit code as an argument. Without this
change, riscv-vp ignores this exit code and always terminates with a
zero exit status. This is problematic when using Benchmark suites like
embench (for example) which indicate a result verification failure via
this exit code.

This is a bit challenging to implement as we can't directly propagate
an exit status via shall_exit to sc_main. In order to work around this,
this commit calls exit(3) directly for now if the requested exit-code
is non-zero.

This change can introduce breakage when software does not set the a0
register when performing a SYS_exit syscall. I briefly checked our
example code in `sw/` and it seems that all of them set a0 to a zero
value.