sw: prevent endless loop on optimizing compilers, if wfi is removed
Example problem:
```
_Bool rx_available = 0;
...
while (!rx_available) {
	asm volatile ("wfi");
}
```
When wfi is removed (e.g. for experiments), the reading of "rx_available"
is removed by an optimizing compiler, resulting in an infinite loop.

Solution:
Make such variables volatile so that reads are never optimized out.

This patch fixes "basic-dma", "blocking-sleep" and "simple-sensor".
("periphery-in-the-loop" uses the same pattern, but the variable is
already defined volatile.)

--
manfred.schlaegl@jku.at
Institute for Complex Systems, JKU Linz
https://ics.jku.at/