GUI-VP: VNCSimpleFB: eliminate intermediate frame buffer (performance)
By matching the pixel format of the VNCServer with that of Linux
simple-framebuffer we not only can eliminate the format conversion,
but also the immediate frame buffer which increases overall performance.

With this, we also disable area tracking, since it is more costly to the
simulation that a full update (which is done in the rfb thread anyways).