GUI-VP: Linux: Added memory-mapped file to support a rootfs via linux physmap drivers
Having the rootfs as initramfs (payload in the kernel/opensbi image)
leads to two disadvantages:
 1. System start-up time scales with image size (content of whole image
    is copied on boot)
 2. initramfs takes up space in main memory

With this, we add support for a memory mapped file which can be used to
hold a root file system image by using the physmap drivers of the
linux kernels mtd subsystem.
For RV64 the file size is set to 512MB by default. For RV32 the size is
limited to 64MB due to the very limited vmalloc size.

A rootfs image file can be mapped by the VP to the specific memory location.
The kernel can then use this memory area as memory mapped block device and
is able to mount the rootfs from this.

Using this mechanism can lead to much faster system start-up times (~1/2
time to login compared to initramfs) and more free main memory.