GUI-VP: Linux: Added memory-mapped file to support persistant storage (data) via linux physmap drivers
Similar as we did for rootfs we add a second memory mapped file for
persistent storage (data).

With this, we can realize a very common pattern in embedded system
design: A read-only root filesystem to hold system and constant data,
and a read-write data file for persistent / generated data.

The reason for using two separate mem-mapped files here instead of one
including partitions, is simplicity. It is much easier for a user to
handle two files for root and data, instead of one file including
partitions.