ISS: fixed EBREAK instruction (trap, semihosting)
Before this, the implementation of EBREAK contained only breakpoint
handling for debug mode. A corresponding trap was not issued.

However, riscv-spec-20191213 (page 27) states:
"EBREAK was primarily designed to be used by a debugger to cause
execution to stop and fall back into the debugger. EBREAK is also
used by the standard gcc compiler to mark code paths that should
not be executed.
Another use of EBREAK is to support “semihosting”, where the
execution environment includes a debugger that can provide services
over an alternate system call interface built around the EBREAK
instruction.
"

To support the latter two use-cases, EBREAK now triggers a trap with
exception code 3, according to riscv-privileged-20211203 (Table 3.7).
(This also fixes a crash of opensbi 1.3 (generic platform,
semihosting check))

To keep previous behaviour of debug mode, we currently do not trigger
the trap if we are in debug mode. This has to be revisited and checked
in the future (see corresponding TODO).

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>