vp: core: fixed support of RISC-V Compressed C instructions on systems using virtual memory
We have support for RISC-V Compressed C instructions.
As a consequence we also might have misaligned 32 bit fetches.

This is a problem if
 1. the fetch happens at a page boundary AND
 2. virtual memory management is used AND
 3. the two involved pages are not stored sequentially in physical memory.

e.g.
<pc>	<instr>		<pc_increment>
0x1FF8	fmul ...	4
0x1FFC	c.lw ...	2 (compressed)
0x1FFE	fmax ...	4		<-- not 32 bit aligned @ page boundary
0x2002	fmin ...	4		<-- not 32 bit aligned

We fix this, by splitting the 32 bit fetch to two 16 bit fetches on 4 KiB
page boundaries.

TODO: Maybe it would be more realistic to split all misaligned 32 bit
accesses in two 16 bit (-> no misaligned accesses on the bus) -> Future work

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>