vp: core: harmonize misa checks for RV32 and RV64
Checks as to whether instructions are valid for a given set of activated
RISC-V extensions (misa csr) (e.g. trap for float load if F not in misa)
were implemented in RV32, but missing in RV64.

With this change we generalise and move the check to the decoder:
 1. The checks are now active for both, RV32 and RV64. are now active
 2. There are now fewer differences between the two ISSs

The disadvantage is that the performance of the RV64 is now somewhat reduced
due to the additional checks. However, a cache for decoded instructions is
to be introduced in a later change. This means that the checks will generally
only be carried out infrequently in the future.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>