vp: core: nuclei_core/(+rv32/rv64): Fixed critical CSR bug in nuclei_core (GD32)
This bug was introduced while porting GD32 to RISC-V VP++.

The original implementation relied on polymorphism for the
csr_table.
The nuclei core has its own CSRs in addition to the default rv32 core
(core/vp/rv32/iss). The original implementation of nuclei:
 * kept the original csr_table in the ISS
 * derived the nuclei ISS from the original ISS and added the extended
   csr_table
 * added a virtual method to get the ISS instance's csr_table
 * modified the nuclei AND the original ISS to access all csrs by
   using this virtual method

The bug in the port to RISC-V VP++ was, that the last point was left
out for the original ISS. With this, two copies of the csr_table
were actively used, one by the parent class (original ISS) and one
by the child class (nuclei ISS). Changes on one where not visible
to the other. This lead for example to non working interrupts:
When firmware tried to set mstatus.mie, the set was applied by the
parent in its csr_table. In the further execution, the child checked
the flag in its csr_table where it was not set. -> Interrupts never
worked.

A naive solution would be to add the missing part, the access of csrs
via the virtual method in the parent ISS. However, this comes at a huge
performance cost, since the ISS accesses CSRS very frequently and
additional redirection for virtual methods is costly (dynamic dispatch).
Another solution would be to just duplicate the code of the parent for
the nuclei ISS. However, this approach is not chosen for obvious reasons
(we already have enough duplication between RV32 and RV64 ;-) )
This patch implements the third approach: Instead of code duplication,
we duplicate the compiled instances by using C++ template classes.

The original ISS is modified to a template with the concrete csr_table
type as parameter. Based on this template, we instruct the compiler to
compile the code for both the "classic" RV32 ISS and the nuclei core ISS.
For consistency, we apply the same approach also to RV64, however we
disable polymorphism for RV64 for now - for more details, please take a
look at "NOTE RVxx.1" in core/rv32/iss.h and core/rv64/iss.h.

The modifications are quite extensive, but this will pay off in terms
of performance in the long run.
A quick performance check (linux boot, dhrystone, prboom) showed that
the modifications made did not cause any significant change in terms of
performance.

Future work: We still have virtual methods in the RV32 ISS. In the
long run, we plan to get rid of polymorphism here all together by using
a C++ template based approach. (something as described in
http://groups.di.unipi.it/~nids/docs/templates_vs_inheritance.html)

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>