vp: core: nuclei_core/rv32(/rv64): generalized run_step (preparation for DBBCache + optimizations)
Generalize run_step by factoring out the interrupt and trap handling.
This is needed for coming up optimizations (DBBCache, fast/slow path).

For RV32 we use virtual methods to support rv32 and the derived
nuclei_core (GD32).
This comes with additional cost in run_step esp. due to the dynamic dispatch
of handle_interrupt, which is called for each instruction. However, this
is only temporary. As soon as we add DBBCache and other optimizations we
will get this back.

For RV64 we do structurally the same (harmonization with RV32). However,
we don't declare the methods virtual yet i.e. there is no performance impact
(see "NOTE RVxx.1" in core/rv64/iss.h for more details).

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>