vp: core: rv32/rv64: rework to C-style templates (preparation for DBBCache + optimizations)
In the previous commit
"vp: core: nuclei_core/(+rv32/rv64): Fixed critical CSR bug in nuclei_core (GD32)"
we reworked the RV32 and RV64 ISS using C++ templates to fix the derived
nuclei_core (GD32):
 * We converted the ISS to a template
 * For RV32 we created two concrete instances:
   * "ISS" which uses the classic RV32 csr_table for classic RV32 platforms,
     and
   * "NUCLEI_ISS_BASE" which uses the nuclei_csr_table as base class for the
     derived nuclei_core
   Since the latter makes use of polymorphism, we have to use virtual
   methods which come at a runtime cost (dynamic dispatch) which also
   affected the classic RV32 ISS
 * For RV64 we created only one concrete instance, namely the "ISS" which uses
   the classic RV64 csr_table, since we don't have derived cores yet. However,
   we wanted to keep the differences between RV32 and RV64 minimal.
   For RV64 we set the ISS to final and all methods to non-virtual for
   now. So here we introduced no additional cost at runtime.
(see also NOTE RVxx.1 in the source)

Coming to the current change:
For the coming-up integration of DBBCache and built-upon optimizations
we need support for compiler attributes (we need to deal with section
attributes for computed goto / jump labels). However, at time of writing
the GCC compiler ignores attributes in C++ templates, which is widely
accepted as a long-standing bug [1][2].
This means, that classic C++ templates are off the table.

With this patch we rework the RV32 and RV64 ISS again to, what we call
the "C-style" template pattern, which makes heavy use of macros and the
C/C++ preprocessor:
 * We use preprocessor macros to parametrize the ISS class definition
   (iss_ctemplate.h) and implementation (iss_ctemplate.cpp)
 * We create configurations for concrete instances by defining these
   preprocessor macros with concrete values.
 * For each configuration, we include the template definitions and
   implementations into our iss.h and iss.cpp which creates the concrete
   definitions and implementations.
 * After each configurations, we reset the configuration.
More details can be found in "NOTE RVxx.02" in
vp/src/core/rv32/iss_ctemplate_handle.h.

Although this pattern seems hacky at first, it does not only solve
the problem with compiler attributes, but also the problem with the
additional performance cost for the classic RV32 ISS introduced in
the commit mentioned above (dynamic dispatch). The preprocessor
macros provide much more control over the template so we are able
to define the necessary methods virtual for the nuclei core instance
and define the same methods non-virtual for the classic RV32 ISS.

[1] https://stackoverflow.com/questions/74358789/section-attribute-within-template
[2] https://gcc.gnu.org/bugzilla/buglist.cgi?bug_status=UNCONFIRMED&bug_status=NEW&bug_status=ASSIGNED&bug_status=SUSPENDED&bug_status=WAITING&bug_status=REOPENED&cf_known_to_fail_type=allwords&cf_known_to_work_type=allwords&query_format=advanced&short_desc=attribute%20template&short_desc_type=allwordssubstr

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>