vp: core: add support for RV64E
We already had support for RV32E. RV64E was not specified in the past,
but is now - see for example
"The RISC-V Instruction Set Manual Volume I: Unprivileged Architecture, Version 20240411", Chapter 3.

This adds support for RV64E similar as it was implemented for RV32,
which eliminates yet another difference between the RV32 and RV64 ISSes.
Platforms are adapted accordingly to use the RV32/RV64E when the command line
argument --use-E-base-isa is present.

For future work:
RV32E and RV64E is fully supported. However, the current implementation does
not check for (and trap) if registers x16-x31 are used (non-E
instructions are not forbidden).
-> TODO: integrate this check in the decode - see "NOTE RVCOMMON.01" in
   vp/src/core/common/instr.cpp

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>