vp: platform: common/linux: add support for miscdev incl. a hardware random number generator
This adds a new module miscdev to vp/platform/common, which is intended to
provide small miscellaneous functionality.
Currently, the module implements a hardware random number generator (hwrng)
which provides values from the host hwrng via a 32 bit register at offset 0.

The miscdev module is integrated in the linux-vp to provide a hwrng for
Linux (linux timeriomem_rng driver).

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>