vp: introduce RV_ISA_Config to describe the ISA + make Zfh support configurable via command line
Until now, the CSR misa.extensions was used to determine supported
extensions at runtime. Top-levels could modify to csr to enable/disable
extensions (e.g. E extension).
A problem arose with addition of the Zfh extension, which has no
representing bit in misa.extensions.

With this we introduce the (expandable) ISA configuration class RV_ISA_Config.
Currently, the structure holds the exact same bits as misa.extensions +
a bit to indicate if Zfh is supported.

RV_ISA_Config is created in top-levels according to command line
arguments. The structure is passed to the ISSes, which set the value for
misa.extension accordingly. The ISS passes the structure on to the
DBBCache, which passes it to the instruction decoder. RV_ISA_Config is
then used by the decoder to check if instructions are valid for the
current platform.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>