vp: platform: common: memory: reworked error handling and messages
If a non-existing file was provided e.g. via the linux-vp --dtb-file
parameter, the vp silently crashed with a segmentation fault making it
hard to track down the issue.
The reason for the fault is, that mapped_file_source used in
SimpleMemory::load_binary_file expects the file to exist (see
documentation).

With this, we add an additional check before mapped_file_source. We open
the file read-only to check for existence and read permissions. If this
fails we print a error message and abort.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>