vp: core: refactor/work decoder: fixed double-used term opcode, improved typesafety, cleanups
The refactoring mainly considers the double-used term "opcode": "Opcode" was
used in the VP to refer to the internal operation id generated by the
decoder and used by the ISS to select instruction implementations. However,
"opcode" is also a field in RISC-V instruction encoding.
In this patch, we fix this double-use by renaming the former "Opcode" to
"OpId".

The patch additionally includes:
 * Various cleanups and refactorings related to the above change
 * Namespace cleanups (Operation vs. Instruction)
 * Improved type safety through explicit specification of the enum type for
   enum value references
 * Differentiation of uncompressed and compressed operation ids (OpId
   vs. C_OpId) - This also fixes clang compilation issues
 * Removal of unused/duplicate functionality (regnamePrettyStr -> see
   regfile.h)

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>