vp: platform: redesign of uart/slip/fu540_uart to improve re-use
Before this change, the actual hardware model was on top a C++ class
inheritance hierarchy. The uart (console output) and slip (networking)
where the leafs of this hierarchy.
This architecture made it exceptionally hard to re-use functionality for
new uart hardware models (e.g. a ns16550a).

With this, we completely reworked the design.
 * All the threaded RX/TX fifo handling was moved to the abstract
   channel_if class
 * Console and slip functionality was moved to specializations from
   channel_if, namely to channel_console and channel_slip
 * The fu540_uart hardware model takes a concrete channel_fi instance in
   the constructor.

With this design, it is now possible to create very compact hardware
models of new uart peripherals, that can automatically use the
functionality provided by the channels.

NOTE: This change temporarily breaks the forwarding of UART0/1 TX/RX
to the virtual breadboard on the hifive platform (--forward-uart0-to-vbb
and --forward-uart1-to-vbb). The tunnel-uart must be reworked to
a channel_tunnel (derived from Channel_IF) to solve this.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>