vp: platform: second part of the uart/channel redesign (fixes hifive uart breadboard)
This patch is a sequel to
763fb15c785b6317961ac250ec5fd44c4681170f
vp: platform: redesign of uart/slip/fu540_uart to improve re-use

This patch re-enables the support for forwarding of UART0/1 TX/RX to the
virtual breadboard on the hifive platform (--forward-uart0-to-vbb
and --forward-uart1-to-vbb).
The former Channel_IF is further separated in
 * Channel_IF:
   * basic interface for uarts
   * only handles rx/tx fifos
   * basis for the newly introduced channel_tunnel (hifive)
 * Channel_FD_IF:
   * same functionality as the former Channel_IF
   * derived from Channel_IF
   * adds thread based file descriptor (FD) based back-end communication
   * basis for channel_console and channel_slip

Based on this, tunnel-uart (hifive platform) is redesigned to channel_tunnel,
derived from Channel_IF.
Finally, support for UART0/1 TX/RX forwarding to the virtual breadboard is re-enabled
in the hifive platform.

The patch also contains some minor improvements, fixes and further
cleanups.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>