vp: core: rv32: add missing csr mstatush and missing bits (ube, sbe, mbe)
according to privileged spec
"The MBE, SBE, and UBE bits in mstatus and mstatush are WARL fields that
control the endianness of memory accesses other than instruction fetches.
Instruction fetches are always little-endian."

The mstatush csr was completely missing and accesses lead to a trap (violation
of the privileged spec). Interestingly, this was never a problem in our use
cases until we switched to OpenSBI version 1.6 in the GUI-VP Kit.
OpenSBI introduced the access to mstatush with commit
"lib: sbi: add Smdbltrp ISA extension support"
c46a937fd9071660979ff68e5c6310e62c9d1770
(https://github.com/riscv-software-src/opensbi.git)

With this we
 1. add mstatush, including the MBE and SBE bits, and
 2. add the UBE bit in mstatus
according to the privileged spec.
All bits are 0 (little-endian) and WARL (read-only).

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>