vp: platform: common: sifive_spi: remove incorrect alternative fifo full handling
According to the RTL model
(https://github.com/sifive/sifive-blocks/tree/master), the txfifo puts
backpressure on the rxfifo -> If the rxfifo is full, writes to txdata
are ignored. However(!), the full bit in txdata is never set in this case.
Hence, the alternative implementation is incorrect and therefore removed.

Please note, that the alternative implementation was disabled by
default, i.e., this change does not modify the default behavior.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>