vp: Revision of include handling to avoid conflicts between vp and system header files
Co-Author: Jonas REICHHARDT <reichhardt.jonas@gmail.com>

This is an overall cleanup of the VP's include file handling and fixes
for example C++17 to C++20/C++23 migration problems.

Until now, the VP heavily used CMake's "target_include_directories"
to add project subdirectories to the compiler's directory search list for
header files. (behind the scene, cmake uses the compiler's "-I" option [1]
to realize this). With this, it was for example possible to just write
 #include <bus.h>
instead of
 #include "platform/common/bus.h"
because "platform/common" was in the search list ("-Iplatform/common").

Although, this seems to add some convenience at first glance,
it comes with a major pitfall, namely the danger of conflicts between
include files provided by the VP and the system (e.g. standard libraries).

A concrete example of such a conflict becomes apparent when trying to
migrate the VP from C++17 to C++20. The C++20 atomic.h now includes
syscall.h. However, since the VP has a project-specific syscall.h in its
header search paths (e.g., "core/rv64/syscall.h"), the compiler uses the
VP's syscall.h instead of the system's syscall.h, which of course breaks
the compilation.

With this change, we are largely eliminating the use of CMake's
target_include_directories. The includes have been revised to paths
relative to the vp-root directory.

[1]
https://gcc.gnu.org/onlinedocs/gcc/Directory-Options.html#index-I

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>