vp: platform: tiny32/tiny64: top-level generalization (2/2): merge tiny64 in tiny
Generalizes the top-level code for "tiny32" and "tiny64" to "tiny".

We do this in two steps (rename and generalization) to get a clean git
history.

This is the second step: We merge the top-level of "tiny64" in "tiny" and
adapt the CMakeLists to build "tiny32-vp" and "tiny64-vp" from one souce
(similar to "basic", "linux", ...)

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>