vp: platform: tiny32-mc/tiny64-mc: top-level generalization (2/2): merge tiny64-mc in tiny-mc
Generalizes the top-level code for "tiny32-mc" and "tiny64-mc" to "tiny-mc".

We do this in two steps (rename and generalization) to get a clean git
history.

This is the second step: We merge the top-level of "tiny64-mc" in "tiny-mc"
and adapt the CMakeLists to build "tiny32-mc-vp" and "tiny64-mc-vp" from one
souce (similar to "basic", "linux", ...)

Furhtermore, we renamed the binaries to "tiny32-mc" to "tiny32-mc-vp" and
"tiny64-mc" to "tiny64-mc-vp" for consistency with the other vp
platforms.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>