vp: core: common: allow for clean use of register definitions (cleanup)
Before that, we had to define a concrete (e.g. rv32) register file based
on the RegFile_T template only to be able to access constants like
number of registers or register enums.

With this, we factored out all basic/generic definitions for RISC-V register
files to RegFileBase. RegFileBase is used as base class for RegFile_T
and can be used to access generic constants (see syscall.h).

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>