vp: core: dbbcache: fixed random cycle counting bug (e.g. mcycles)
Before this fix, the start-value used for the cycle counter in
dbbcache was not properly initialized which could lead to random start
values (e.g. in mcycles).

Background: In the initialization, we first start with the dummy block.
Then, we use the switch block methods (find_create_block or
switch_block_dummy) to properly set up the real initial block.
These switch methods transfer over the cycles of the last block to the
new block by accessing the tail element (curEntryIdx+1) of the last block.
In the init-case, the start block is also the dummy block and since
curEntryIdx is set to 0, the transfer accesses the second entry of
this dummyBlock which is not properly initialized.

Fix: Set curEntryIdx to -1 in the initialization -> the cycle transfer
in the switch block methods uses the first entry of the dummy block,
which is properly initialized to 0.

Signed-off-by: Manfred SCHLAEGL <manfred.schlaegl@gmx.at>